Init media and playColor in MediaWidget
